+++ forge/core/string_utilities.hpp
@@ -21,5 +21,5 @@
 namespace forge {
 std::string trim(const std::string& s);
 
-uint32_t count_whitespace_chars_at_start(const std::string& s);
+size_t count_whitespace_chars_at_start(const std::string& s);
 }  // namespace forge
